requirements.txt
feedparser
PyYAML
requests
beautifulsoup4
